update migrations to use bigIncrements and unsignedBigInteger methods

+++ database/migrations/2015_07_03_010110_create_posts_table.php
@@ -13,9 +13,9 @@ class CreatePostsTable extends Migration
     public function up()
     {
         Schema::create('posts', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id')->index();
-            $table->unsignedInteger('category_id')->index();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id')->index();
+            $table->unsignedBigInteger('category_id')->index();
             $table->string('title');
             $table->text('body');
             $table->timestamps();
