+++ database/migrations/2015_07_03_011315_create_comments_table.php
@@ -13,9 +13,9 @@ class CreateCommentsTable extends Migration
     public function up()
     {
         Schema::create('comments', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('post_id')->index();
-            $table->unsignedInteger('user_id')->index();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('post_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
             $table->text('body');
             $table->timestamps();
         });
