+++ database/migrations/2017_02_14_225431_create_post_tag_table.php
@@ -14,9 +14,9 @@ class CreatePostTagTable extends Migration
     public function up()
     {
         Schema::create('post_tag', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('post_id')->index();
-            $table->unsignedInteger('tag_id')->index();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('post_id')->index();
+            $table->unsignedBigInteger('tag_id')->index();
 
             $table->foreign('post_id')->references('id')->on('posts')->onDelete('cascade');
             $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
